Serialize navigation load results
