+++ .changeset/popular-lemons-do.md
@@ -1,0 +1,5 @@
+---
+'@quilted/preact-router': patch
+---
+
+Serialize navigation load results
