+++ packages/preact-router/source/components/Navigation.tsx
@@ -1,6 +1,7 @@
 import type {RenderableProps} from 'preact';
 import {useMemo} from 'preact/hooks';
 import {useBrowserDetails} from '@quilted/preact-browser';
+import {useAsyncActionCacheSerialization} from '@quilted/preact-async';
 
 import {Router} from '../Router.ts';
 import {RouterContext} from '../context.ts';
@@ -12,19 +13,25 @@ export interface NavigationProps<Context = unknown> {
   router?: Router;
   routes?: readonly RouteDefinition<any, any, Context>[];
   context?: Context;
+  serialize?: boolean;
 }
 
 export function Navigation<Context = unknown>({
-  router,
+  router: explicitRouter,
   routes,
   context,
   children,
+  serialize = true,
 }: RenderableProps<NavigationProps<Context>>) {
   const browser = useBrowserDetails({optional: true});
-  const resolvedRouter = useMemo(
-    () => router ?? new Router(browser?.request.url),
-    [router],
+  const router = useMemo(
+    () => explicitRouter ?? new Router(browser?.request.url),
+    [explicitRouter],
   );
+
+  if (router.cache && serialize) {
+    useAsyncActionCacheSerialization(router.cache, {name: 'router'});
+  }
 
   const content = routes ? (
     <Routes list={routes} context={context} />
@@ -33,8 +40,6 @@ export function Navigation<Context = unknown>({
   );
 
   return (
-    <RouterContext.Provider value={resolvedRouter}>
-      {content}
-    </RouterContext.Provider>
+    <RouterContext.Provider value={router}>{content}</RouterContext.Provider>
   );
 }
