+++ packages/preact-router/source/index.ts
@@ -1,5 +1,10 @@
 export * from './types.ts';
-export {Router, type RouterOptions, type NavigateOptions} from './Router.ts';
+export {
+  Router,
+  RouterNavigationCache,
+  type RouterOptions,
+  type NavigateOptions,
+} from './Router.ts';
 export {route, fallbackRoute} from './route.ts';
 
 export {Link} from './components/Link.tsx';
